support accessing dashboard with url including credentials
