+++ sky/cli.py
@@ -6248,13 +6248,11 @@ def api_info():
                            name=api_server_user['name'])
     else:
         user = models.User.get_current_user()
-    dashboard_url = server_common.get_dashboard_url(url)
-    click.echo(f'Using SkyPilot API server: {url}\n'
+    click.echo(f'Using SkyPilot API server and dashboard: {url}\n'
                f'{ux_utils.INDENT_SYMBOL}Status: {api_server_info["status"]}, '
                f'commit: {api_server_info["commit"]}, '
                f'version: {api_server_info["version"]}\n'
-               f'{ux_utils.INDENT_SYMBOL}User: {user.name} ({user.id})\n'
-               f'{ux_utils.INDENT_LAST_SYMBOL}Dashboard: {dashboard_url}')
+               f'{ux_utils.INDENT_LAST_SYMBOL}User: {user.name} ({user.id})')
 
 
 @cli.group(cls=_NaturalOrderGroup)
