+++ sky/dashboard/src/data/connectors/clusters.jsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect, useCallback } from 'react';
 import { showToast } from '@/data/connectors/toast';
-import { ENDPOINT } from '@/data/connectors/constants';
+import { apiClient } from '@/data/connectors/client';
 import dashboardCache from '@/lib/cache';
 
 /**
@@ -46,23 +46,11 @@ const clusterStatusMap = {
 
 export async function getClusters({ clusterNames = null } = {}) {
   try {
-    const response = await fetch(`${ENDPOINT}/status`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        cluster_names: clusterNames,
-        all_users: true,
-      }),
+    const clusters = await apiClient.fetch('/status', {
+      cluster_names: clusterNames,
+      all_users: true,
     });
-    // TODO(syang): remove X-Request-ID when v0.10.0 is released.
-    const id =
-      response.headers.get('X-Skypilot-Request-ID') ||
-      response.headers.get('X-Request-ID');
-    const fetchedData = await fetch(`${ENDPOINT}/api/get?request_id=${id}`);
-    const data = await fetchedData.json();
-    const clusters = data.return_value ? JSON.parse(data.return_value) : [];
+
     const clusterData = clusters.map((cluster) => {
       // Use cluster_hash for lookup, assuming it's directly in cluster.cluster_hash
       let region_or_zone = '';
@@ -125,31 +113,18 @@ export async function streamClusterJobLogs({
   workspace,
 }) {
   try {
-    const response = await fetch(`${ENDPOINT}/logs`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        // TODO(hailong): set follow to true?
-        // - Too much streaming requests may consume too much resources in api server
-        // - Need to stop the api request in different cases
+    await apiClient.stream(
+      '/logs',
+      {
         follow: false,
         cluster_name: clusterName,
         job_id: jobId,
         override_skypilot_config: {
           active_workspace: workspace || 'default',
         },
-      }),
-    });
-    // Stream the logs
-    const reader = response.body.getReader();
-    while (true) {
-      const { done, value } = await reader.read();
-      if (done) break;
-      const chunk = new TextDecoder().decode(value);
-      onNewLog(chunk);
-    }
+      },
+      onNewLog
+    );
   } catch (error) {
     console.error('Error in streamClusterJobLogs:', error);
     showToast(`Error in streamClusterJobLogs: ${error.message}`, 'error');
@@ -158,26 +133,14 @@ export async function streamClusterJobLogs({
 
 export async function getClusterJobs({ clusterName, workspace }) {
   try {
-    const response = await fetch(`${ENDPOINT}/queue`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
+    const jobs = await apiClient.fetch('/queue', {
+      cluster_name: clusterName,
+      all_users: true,
+      override_skypilot_config: {
+        active_workspace: workspace,
       },
-      body: JSON.stringify({
-        cluster_name: clusterName,
-        all_users: true,
-        override_skypilot_config: {
-          active_workspace: workspace,
-        },
-      }),
     });
-    // TODO(syang): remove X-Request-ID when v0.10.0 is released.
-    const id =
-      response.headers.get('X-Skypilot-Request-ID') ||
-      response.headers.get('X-Request-ID');
-    const fetchedData = await fetch(`${ENDPOINT}/api/get?request_id=${id}`);
-    const data = await fetchedData.json();
-    const jobs = JSON.parse(data.return_value);
+
     const jobData = jobs.map((job) => {
       let endTime = job.end_at ? job.end_at : Date.now() / 1000;
       let total_duration = 0;
