+++ sky/dashboard/src/pages/clusters.js
@@ -1,6 +1,11 @@
 import React from 'react';
 import Head from 'next/head';
-import { Clusters } from '@/components/clusters';
+import dynamic from 'next/dynamic';
+
+const Clusters = dynamic(
+  () => import('@/components/clusters').then((mod) => mod.Clusters),
+  { ssr: false }
+);
 
 export default function ClustersPage() {
   return (
