+++ sky/dashboard/src/pages/infra/[context].js
@@ -1,6 +1,11 @@
 import React from 'react';
-import { GPUs } from '@/components/infra';
 import Head from 'next/head';
+import dynamic from 'next/dynamic';
+
+const GPUs = dynamic(
+  () => import('@/components/infra').then((mod) => mod.GPUs),
+  { ssr: false }
+);
 
 export default function InfraContextPage() {
   return (
